Adding a node at the start of the list

+++ 0x12-singly_linked_lists/_strlen.c
@@ -4,9 +4,9 @@
  * _strlen - to find the length of a string
  * @str: pointer to a string
  */
-size_t _strlen(const char *s)
+int _strlen(const char *s)
 {
-	size_t len = 0;
+	int len = 0;
 	while (s[len])
 	{
 		len++;
